v0.01
A small web app that brings together some of the things that I’ve
learned from Treehouse and the pain that is Rule of the Road.

At the moment, this is a linear app. You have to start from the
beginning to get to the end. In future updates there will be random
questions generated so that it will test the questions that may come up
in fleet board.

Additional features that need to be added include:
- highlighting the bits that you go wrong
- accepting subtle differences in typing (i.e. spaces where there
aren’t any)
- a score tracker on the outside of the box
- an updated UI so it doesn’t just look like the crap I’ve thrown
together now

Some features I’d like to see:
- beginner, harder and hardest questions